requirements.txt
paho-mqtt==1.6.1
pylint==2.17.7
pyowm==3.3.0
RPi.GPIO==0.7.1
matplotlib==3.5.3
